// src/app/core/model/roles/model/Roles.ts
export interface VMRole {
  id: any;
  name: string;
  description: string;
}
export interface VMCreateRole {
  name: string;
  description: string;
}
export interface VMUpdateRole {
  id: any;
  name: string;
  description: string;
}
export interface VMDeleteRole {
  id: any;
  name: string;
  description: string;
}